Refactor database service structure
Move database.js to src/services/db/ and consolidate character query
functions to accept optional projectId arguments.
