+++ src/pages/SceneDetailPage.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react'
 import { useParams, useNavigate, Link } from 'react-router-dom'
 import { Card, Button, Spinner, TextInput, Label, Select, Modal } from 'flowbite-react'
-import { sceneCrud, shootingDayCrud, getShootingDaysByProject, getCharactersByProject, getCostumesByProject } from '../services/database'
+import { sceneCrud, shootingDayCrud, getShootingDaysByProject, getCharacters, getCostumesByProject } from '../services/db/database'
 import { useProject } from '../contexts/ProjectContext.jsx'
 
 function SceneDetailPage() {
@@ -49,7 +49,7 @@ function SceneDetailPage() {
         if (currentProjectId) {
           const [shootingDaysData, charactersData, costumesData] = await Promise.all([
             getShootingDaysByProject(currentProjectId),
-            getCharactersByProject(currentProjectId),
+            getCharacters(currentProjectId),
             getCostumesByProject(currentProjectId)
           ])
