+++ src/services/db/collections/shootingDays.js
@@ -1,5 +1,5 @@
 // Shooting Day collection operations
-import { createCRUDOperations } from '../utils.js';
+import { createCRUDOperations } from '../database.js';
 import { replicateAppwrite } from 'rxdb/plugins/replication-appwrite';
 
 export const shootingDaySchema = {
